requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.1.0
Django==5.0.2
django-filter==23.5
django-heroku==0.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-spectacular==0.27.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
install==1.3.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
